📦 FEAT: add navbar component + organize pages
