+++ src/components/navbar.tsx
@@ -1,0 +1,57 @@
+const navbar = () => {
+  return (
+    <div className="navbar bg-base-100">
+      <div className="navbar-start">
+        <div className="dropdown">
+          <div tabIndex={0} role="button" className="btn btn-ghost lg:hidden">
+            <svg
+              xmlns="http://www.w3.org/2000/svg"
+              className="h-5 w-5"
+              fill="none"
+              viewBox="0 0 24 24"
+              stroke="currentColor"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                strokeWidth="2"
+                d="M4 6h16M4 12h8m-8 6h16"
+              />
+            </svg>
+          </div>
+          <ul
+            tabIndex={0}
+            className="menu menu-sm dropdown-content mt-3 z-[1] p-2 shadow bg-base-100 rounded-box w-52"
+          >
+            <li>
+              <a href="/pages/about">Main</a>
+            </li>
+            <li>
+              <a href="/pages/cognicy">Cognicy</a>
+            </li>
+            <li>
+              <a href="/pages/developers">Developers</a>
+            </li>
+          </ul>
+        </div>
+        <a className="btn btn-ghost text-xl">daisyUI</a>
+      </div>
+      <div className="navbar-center hidden lg:flex">
+        <ul className="menu menu-horizontal px-1">
+          <li>
+            <a>Main</a>
+          </li>
+          <li>
+            <a>Cognicy</a>
+          </li>
+          <li>
+            <a>Developers</a>
+          </li>
+        </ul>
+      </div>
+      <div className="navbar-end"></div>
+    </div>
+  );
+};
+
+export default navbar;
